requirements.txt
streamlit>=1.35
openai>=1.30
chromadb>=0.5
rank-bm25>=0.2.2
wikipedia>=1.4.0
tiktoken>=0.7
ruff>=0.4
pytest>=8.0
python-dotenv>=1.0
